@@ -21,20 +21,17 @@ setup(
         "Operating System :: OS Independent",
         "Programming Language :: Python",
         "Programming Language :: Python :: 3",
-        "Programming Language :: Python :: 3.9",
-        "Programming Language :: Python :: 3.10",
-        "Programming Language :: Python :: 3.11",
         "Programming Language :: Python :: 3.12",
         "Programming Language :: Python :: 3.13",
+        "Programming Language :: Python :: 3.14",
         "Programming Language :: Python :: Implementation :: CPython",
-        "Programming Language :: Python :: Implementation :: PyPy",
         "Topic :: System :: Logging",
     ],
     name="logxpy",
     version=versioneer.get_version(),
     cmdclass=versioneer.get_cmdclass(),
-    description="Logging library that tells you why it happened",
-    python_requires=">=3.9.0",
+    description="Modern structured logging library with hierarchical Sqid task IDs",
+    python_requires=">=3.12.0",
     install_requires=[
         # Internal code documentation:
         "zope.interface",
