@@ -2,14 +2,14 @@
 13_complex_data_pipeline.py - Complex Data Processing Pipeline
 Demonstrates 7-level nested structure with ETL operations and microservices
 """
-
+from pathlib import Path
 import asyncio
-import sys
 import time
 
 from logxpy import aaction, log, to_file
 
-to_file(sys.stdout)
+LOG_FILE = Path(__file__).with_suffix(".log")
+to_file(open(LOG_FILE, "w"))
 
 
 # ============================================================================
